requirements.txt
djangorestframework==3.15.2
drf-spectacular==0.27.2
django-cors-headers==4.4.0
ruff==0.5.7
mypy==1.11.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pre-commit==3.8.0
pytest==8.3.2
pytest-django==4.8.0
djangorestframework-simplejwt==5.3.1
rest-framework-simplejwt==0.0.2
djangorestframework-camel-case==1.4.2
drf-orjson-renderer==1.7.2
django-filter==24.3